Rename sk_class.rs
